#19 refactor(input):테스트코드 제거/컨벤션 참고하여  수정
